+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 3. Support the [`<path>`](https://svgwg.org/svg2-draft/paths.html#PathElement) element
 4. Basic [`paint`](https://svgwg.org/svg2-draft/painting.html#SpecifyingPaint) support
 5. Shape [stroke](https://svgwg.org/svg2-draft/painting.html#StrokeProperties) support
+6. [`viewBox`](https://svgwg.org/svg2-draft/coords.html#ViewBoxAttribute) support on the [`<svg>`](https://svgwg.org/svg2-draft/struct.html#SVGElement) element
 
 ### Performance improvements
 1. `ColorMaterialRenderer` and `TextureMaterialRenderer` are now globally shared instead of per `Artwork`, which previously made the creation time of tiles significantly slower.